+++ src/lib/Components/Inbox/Bids/ActiveBid.tsx
@@ -8,7 +8,10 @@ import { BodyText, MetadataText } from "../Typography"
 import colors from "../../../../data/colors"
 import OpaqueImageView from "../../OpaqueImageView"
 
-const Container = styled.View`margin: 17px 20px 0px;`
+const Container = styled.View`
+  margin: 17px 20px 0px;
+  height: 80px;
+`
 
 const Content = styled.View`
   flex: 1;
@@ -56,7 +59,7 @@ interface State {
   status: BidStatus
 }
 
-class ActiveBid extends React.Component<any, State> {
+class ActiveBid extends React.Component<RelayProps, State> {
   constructor(props) {
     super(props)
 
@@ -168,3 +171,43 @@ export default Relay.createContainer(ActiveBid, {
     `,
   },
 })
+
+interface RelayProps {
+  bid: {
+    is_leading_bidder: boolean | null
+    active_bid: {
+      id: string
+      max_bid: {
+        cents: number | null
+        display: string | null
+      } | null
+      sale_artwork: {
+        lot_label: string | null
+        lot_number: string | null
+        position: number | null
+        reserve_status: string | null
+        counts: {
+          bidder_positions: boolean | number | string | null
+        } | null
+        sale: {
+          live_start_at: string | null
+          end_at: string | null
+          is_live_open: boolean | null
+          is_closed: boolean | null
+        } | null
+        highest_bid: {
+          cents: number | null
+          display: string | null
+        } | null
+        artwork: {
+          id: string
+          title: string | null
+          image: {
+            url: string | null
+          } | null
+          artist_names: string | null
+        } | null
+      } | null
+    } | null
+  }
+}