+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 -   [dev] Allows toggling the back button by pressing space - orta
 
 ###### Messaging
-
+-   Fixes pagination for Conversations - luc
+-   Adds pull to refresh support to Inbox view - luc
 -   Adds an `ImagePreview` to each Conversation, Relay-ified messages - matt
 -   Adds snapshot tests for active bids - luc
 -   Update status label colors - luc